refactor: migrate all packages to ESM
