+++ packages/generators/package.json
@@ -2,7 +2,8 @@
   "name": "@webpack-cli/generators",
   "version": "2.5.0",
   "description": "Webpack-CLI generators",
-  "main": "lib/index.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "types": "lib/index.d.ts",
   "keywords": [],
   "license": "MIT",
