+++ packages/info/package.json
@@ -2,7 +2,8 @@
   "name": "@webpack-cli/info",
   "version": "1.5.0",
   "description": "Outputs info about system and webpack config",
-  "main": "lib/index.js",
+  "type": "module",
+  "exports": "./lib/index.js",
   "types": "lib/index.d.ts",
   "license": "MIT",
   "engines": {
