+++ packages/serve/src/index.ts
@@ -1,5 +1,5 @@
 import type { Compiler, cli } from "webpack";
-import { IWebpackCLI, WebpackDevServerOptions } from "webpack-cli";
+import { IWebpackCLI, WebpackDevServerOptions } from "webpack-cli/src";
 
 const WEBPACK_PACKAGE = process.env.WEBPACK_PACKAGE || "webpack";
 const WEBPACK_DEV_SERVER_PACKAGE = process.env.WEBPACK_DEV_SERVER_PACKAGE || "webpack-dev-server";
