+++ tsconfig.json
@@ -4,7 +4,7 @@
   "compilerOptions": {
     "skipLibCheck": true,
     "target": "es2017",
-    "module": "commonjs",
+    "module": "ES6",
     "lib": ["es6", "es2017"],
     "strict": true,
     "rootDir": ".",
